Add files via upload
The pot_file.zip contains the potential file e_1400_1800.dat  (The first column is theta in radians, the second column is Z in angstrom and the third column is the potential in Hartree. This "e_1400_1800.dat " needs to be extracted first to run the c++ code to generate the 2D eigenfunctions.